refactor: split profile sections into components
